requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
python-dateutil==2.8.2
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
requests==2.31.0
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0
